feat: :construction: Work on the mobile menu is ongoing.
I was thinking about using JS to make the menu interactive, but I couldn't figure it out. So, I looked for a different way to do it. I've found the solution, and now what's left is to style the dropdown menu and make some adjustments in the CSS to finish the project.