feat: :sparkles: The second part of the web project has been completed.
It is worth noting that since the project does not have JavaScript, to view the last two HTML pages, you need to select the respective button for the empty cart and the first button of the first product for the purchase one. To return to the cart menu, you need to go back to any of the other HTML pages and select the cart option again.